Presentation: Split full stack test suites (#6533)
